+++ lib/src/modules/common/constants.dart
@@ -1,4 +1,5 @@
 import 'package:github/github.dart';
+import 'package:sidekick/generated/l10n.dart';
 
 /// Sidekick title
 const kAppTitle = 'Sidekick';
@@ -30,17 +31,17 @@ const kFlutterTagsUrl = 'https://github.com/flutter/flutter/releases/tag/';
 /// Description for the channels
 Map<String, String> channelDescriptions = {
   'stable':
-      '''We recommend that you use this channel for all production app releases. Roughly once a quarter, a branch that has been stabilized on beta will become our next stable branch and we will create a stable release from that branch.''',
+      S.current.stableChannelDescription,
   'beta':
-      '''Branch created from master for a new beta release at the beginning of the month, usually the first Monday. This will include a branch for Dart, the Engine and the Framework.''',
+      S.current.betaChannelDescription,
   'dev':
-      '''The latest fully-tested build. Usually functional, but see Bad Builds for a list of known "bad" dev builds.''',
+      S.current.devChannelDescription,
   'master':
-      '''The current tip-of-tree, absolute latest cutting edge build. Usually functional, though sometimes we accidentally break things.''',
+      S.current.masterChannelDescription,
 };
 
 /// Sidekick repository slug
 final kSidekickRepoSlug = RepositorySlug(
-  "leoafarias",
-  "sidekick",
+  'leoafarias',
+  'sidekick',
 );