+++ lib/src/modules/fvm/components/fvm_release_list_item.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
+import 'package:sidekick/generated/l10n.dart';
 
 import '../../../components/atoms/typography.dart';
 import '../../../components/organisms/global_info_dialog.dart';
@@ -33,7 +34,7 @@ class FvmReleaseListItem extends StatelessWidget {
           const SizedBox(width: 20),
           release.isGlobal
               ? ActionChip(
-                  label: const Caption('Global'),
+                  label: Caption(S.of(context).global),
                   avatar: const Icon(MdiIcons.information, size: 20),
                   onPressed: () {
                     showGlobalInfoDialog(context);