+++ lib/src/modules/search/components/search_bar.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:material_floating_search_bar/material_floating_search_bar.dart';
+import 'package:sidekick/generated/l10n.dart';
 
 import '../../../hooks/floating_search_bar_controller.dart';
 import '../../../providers/navigation_provider.dart';
@@ -66,7 +67,7 @@ class SearchBar extends HookWidget {
               ),
             ),
             FloatingSearchBar(
-              hint: 'Search... ',
+              hint: S.of(context).search,
               controller: controller,
               progress: isLoading.value,
               shadowColor: Colors.transparent,
@@ -99,9 +100,9 @@ class SearchBar extends HookWidget {
                   return const SizedBox(height: 0);
                 }
                 if (results.isEmpty) {
-                  return const Padding(
-                    padding: EdgeInsets.all(20),
-                    child: Text('No Results'),
+                  return Padding(
+                    padding: const EdgeInsets.all(20),
+                    child: Text(S.of(context).noResults),
                   );
                 }