+++ CHANGELOG.md
@@ -1,6 +1,31 @@
 # Changelog
 
-## Unreleased
+## 0.1.1 - 2021-06-19
+
+### Added
+
+- Github trending repos
+- Flutter favorites
+
+## 0.1.0 - 2021-06-16
+
+### Fixed
+
+- Issue when theme brightness does not match platform brightness
+
+## 0.0.9 - 2021-06-15
+
+### Changed
+
+- Updated releases endpoint
+
+## 0.0.8 - 2021-06-03
+
+### Fixed
+
+- master channel not showing in installed releases
+
+## 0.0.7 - 2021-06-02
 
 ### Changed